+++ datafusion/core/src/physical_plan/aggregates/mod.rs
@@ -49,7 +49,6 @@ use std::sync::Arc;
 mod bounded_aggregate_stream;
 mod no_grouping;
 mod row_hash;
-mod row_hash2;
 mod utils;
 
 pub use datafusion_expr::AggregateFunction;
@@ -59,7 +58,6 @@ use datafusion_physical_expr::utils::{
     get_finer_ordering, ordering_satisfy_requirement_concrete,
 };
 
-use self::row_hash2::GroupedHashAggregateStream2;
 use super::DisplayAs;
 
 /// Hash aggregate modes
@@ -214,7 +212,6 @@ impl PartialEq for PhysicalGroupBy {
 enum StreamType {
     AggregateStream(AggregateStream),
     GroupedHashAggregateStream(GroupedHashAggregateStream),
-    GroupedHashAggregateStream2(GroupedHashAggregateStream2),
     BoundedAggregate(BoundedAggregateStream),
 }
 
@@ -223,7 +220,6 @@ impl From<StreamType> for SendableRecordBatchStream {
         match stream {
             StreamType::AggregateStream(stream) => Box::pin(stream),
             StreamType::GroupedHashAggregateStream(stream) => Box::pin(stream),
-            StreamType::GroupedHashAggregateStream2(stream) => Box::pin(stream),
             StreamType::BoundedAggregate(stream) => Box::pin(stream),
         }
     }
@@ -731,22 +727,11 @@ impl AggregateExec {
                 partition,
                 aggregation_ordering,
             )?))
-        } else if self.use_poc_group_by() {
-            Ok(StreamType::GroupedHashAggregateStream2(
-                GroupedHashAggregateStream2::new(self, context, partition)?,
-            ))
         } else {
             Ok(StreamType::GroupedHashAggregateStream(
                 GroupedHashAggregateStream::new(self, context, partition)?,
             ))
         }
-    }
-
-    /// Returns true if we should use the POC group by stream
-    /// TODO: check for actually supported aggregates, etc
-    fn use_poc_group_by(&self) -> bool {
-        //info!("AAL Checking POC group by: {self:#?}");
-        true
     }
 }
 
@@ -1816,10 +1801,10 @@ mod tests {
                     assert!(matches!(stream, StreamType::AggregateStream(_)));
                 }
                 1 => {
-                    assert!(matches!(stream, StreamType::GroupedHashAggregateStream2(_)));
+                    assert!(matches!(stream, StreamType::GroupedHashAggregateStream(_)));
                 }
                 2 => {
-                    assert!(matches!(stream, StreamType::GroupedHashAggregateStream2(_)));
+                    assert!(matches!(stream, StreamType::GroupedHashAggregateStream(_)));
                 }
                 _ => panic!("Unknown version: {version}"),
             }
